add some debug info

+++ include/tgfx/gpu/Context.h
@@ -170,9 +170,9 @@ class Context {
    */
   bool flushAndSubmit(bool syncCpu = false);
 
-  int32_t numDrawCalls() const;
+  size_t numDrawCalls() const;
 
-  int32_t numRenderTasks() const;
+  size_t numRenderTasks() const;
 
   GlobalCache* globalCache() const {
     return _globalCache;
