+++ src/gpu/Context.cpp
@@ -110,11 +110,11 @@ bool Context::flushAndSubmit(bool syncCpu) {
   return result;
 }
 
-int32_t Context::numDrawCalls() const {
+size_t Context::numDrawCalls() const {
   return _drawingManager->numDrawCalls();
 }
 
-int32_t Context::numRenderTasks() const {
+size_t Context::numRenderTasks() const {
   return _drawingManager->numRenderTasks();
 }
